backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
langchain==0.1.20
langchain-community==0.0.38
sentence-transformers==2.7.0
faiss-cpu==1.8.0
ollama==0.2.1
Pillow==10.3.0
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
playwright==1.44.0
beautifulsoup4==4.12.3